requirements.txt
fastapi==0.103.1
python_whois==0.8.0
Requests==2.31.0
seleniumbase==4.18.8
urllib3==1.26.16
uvicorn==0.23.2
